==> notebooks/03_import_infection_data.ipynb <==
# Load and clean Zika infection data
Though the locations of Zika outbreaks were previously imported, the actual infection data was never imported. 

import pandas as pd
import numpy as np
import re
import time
from glob import glob
from csv_pkl_sql import save_it

import matplotlib.pyplot as plt

%matplotlib inline

## Import infection data

data_file_locations = glob('../zika/*/*/data/*.csv')
data = pd.concat([pd.read_csv(x)
                            for x in data_file_locations], axis=0).reset_index(drop=True)

data.drop(['time_period','time_period_type'], axis=1, inplace=True)

data['report_date'] = data.report_date.str.replace('_','-')       
data['report_date'] = pd.to_datetime(data.report_date)

#save_it(data, '03_infection_data_initial_import')

data.head(1)

data.shape

## Clean the data

# Drop municipalities
data = data.loc[data.unit!='municipalities']

data = (data[['report_date', 'location', 'value', 'data_field']]
             .rename(columns={'report_date':'date','value':'zika_cases'}))

for x in data.zika_cases.iteritems():
    try:
        float(x[1])
    except:
        print(x)

data.loc[2414, 'zika_cases'] = 0
data.loc[2783, 'zika_cases'] = 0
data.loc[5192, 'zika_cases'] = 0
data['zika_cases'] = data.zika_cases.fillna(0)
data['zika_cases'] = data.zika_cases.astype(int)

data.query("zika_cases>0").shape, data.shape

# Remove data that don't appear to be directly associated with Zika
excluded_fields = ['cumulative_cases_discarded',
'microcephaly_not',
'gbs_reported',
'zika_not',
'confirmed_acute_fever',
'confirmed_arthralgia',
'confirmed_arthritis', 
'confirmed_rash', 
'confirmed_conjunctivitis',
'confirmed_eyepain', 
'confirmed_headache', 
'confirmed_malaise',
'zika_reported_travel',
'yearly_reported_travel_cases']

mask = data.data_field.isin(excluded_fields)
print(mask.sum(), data.loc[mask, 'zika_cases'].sum(), data.zika_cases.sum())

data = data.loc[mask.pipe(np.invert)]

data.shape[0], data.zika_cases.sum()

data.dtypes

save_it(data, '03_infection_data_final')
