==> notebooks/05_import_mosquito_sighting_info.ipynb <==
# Import sightings of Aedes aegypti mosquitos

import pandas as pd
import numpy as np
import re

from csv_pkl_sql import save_it, pkl_it

## Load the mosquito data

mosquito = pd.read_csv('../aegypti_albopictus/aegypti_albopictus.csv')

mosquito.columns = [x.lower() for x in mosquito.columns]

# according to the code book, x is longitude and y is latitude
mosquito.rename(columns={'x':'longitude', 'y':'latitude'}, inplace=True)

mosquito_clean = pd.concat([mosquito.loc[mosquito.year=='2006-2008'].assign(year=x) 
                      for x in ['2006', '2007', '2008']])

mosquito_clean = pd.concat([mosquito.loc[mosquito.year!='2006-2008'], mosquito_clean])
mosquito_clean = mosquito_clean.loc[mosquito_clean.year.notnull()]
mosquito_clean['year'] = mosquito_clean.year.astype(int)
mosquito_clean.sort_values('year',inplace=True)

mosquito_clean = mosquito_clean.loc[mosquito_clean.year>=2006]

## Compare to airport namings

airport = pd.read_pickle('../pkl/02_airport_information_fallingrain.pkl')

airport_list = airport.country.unique()
mosquito_list = mosquito_clean.country.unique()

np.in1d(airport_list, mosquito_list)

airport_list[-1], [x for x in airport_list if x not in mosquito_list]

mask = mosquito_clean.country=='United States of America'
mosquito_clean.loc[mask,'country'] = 'United States'

sorted(mosquito_clean.country.unique())

airport_list = airport.country.unique()
mosquito_list = mosquito_clean.country.unique()

np.in1d(airport_list, mosquito_list)

save_it(mosquito_clean, '05_mosquito_sightings')
