==> notebooks/07_feature_engineering_and_cleaning_for_MVP.ipynb <==
import pandas as pd
import numpy as np
import dill
from datetime import timedelta
from csv_pkl_sql import save_it, csv_it, pkl_it

import matplotlib.pyplot as plt
import seaborn as sns

%matplotlib inline

## Location data

with open('../pkl/00_cleaned_city_names.pkl', 'r') as fh:
    location_key = dill.load(fh)
location_key.head(1)

with open('../pkl/01_latitude_longitude_google.pkl', 'r') as fh:
    lat_long = dill.load(fh)
lat_long.head(1)

location_key.shape, lat_long.shape

location = pd.merge(location_key, lat_long, on='location', how='inner')
location.head(1)

location.shape

## Airport information

with open('../pkl/02_airport_information_fallingrain.pkl', 'r') as fh:
    airport = dill.load(fh)
airport.head(1)

with open('../pkl/04_merged_latitude_longitude_airport_checkpoint.pkl', 'r') as fh:
    airport2 = dill.load(fh)
airport2.head(1)

Create a dataframe of distance to airport and distance to major airport.x

airport.shape, airport2.shape

airport.kind.unique()

# TODO--convert this to an actual distance

# Closest medium or large airport
airport_coords = airport[['latitude', 'longitude']].values[np.newaxis, :]
places_coords = np.rollaxis(lat_long[['latitude','longitude']].values[np.newaxis, :], 0, -1)
dist_coords = ((places_coords - airport_coords)**2).sum(axis=-1)
min_dist = dist_coords.min(axis=1)

airport_distance = lat_long[['location']].copy()
airport_distance['airport_dist_any'] = min_dist

# Closest large airport
airport_coords = airport.loc[airport.kind=='Large', 
                             ['latitude', 'longitude']].values[np.newaxis, :]
places_coords = np.rollaxis(lat_long[['latitude','longitude']].values[np.newaxis, :], 0, -1)
dist_coords = ((places_coords - airport_coords)**2).sum(axis=-1)
min_dist = dist_coords.min(axis=1)

airport_distance['airport_dist_large'] = min_dist

airport_distance.head()

## Weather information

with open('../pkl/04_weekly_weather.pkl', 'r') as fh:
    weather = dill.load(fh)

weather.head(2)

## Mosquito sightings

with open('../pkl/05_mosquito_sightings.pkl', 'r') as fh:
    mosquito = dill.load(fh)
mosquito.head(1)

# Closest mosquito sighting
mosquito_coords = mosquito[['latitude', 'longitude']].values[np.newaxis, :]
places_coords = np.rollaxis(lat_long[['latitude','longitude']].values[np.newaxis, :], 0, -1)
dist_coords = ((places_coords - mosquito_coords)**2).sum(axis=-1)
min_dist = dist_coords.min(axis=1)

mosquito_distance = lat_long[['location']].copy()
mosquito_distance['mosquito_dist'] = min_dist

mosquito_distance.head()

## Population density

with open('../pkl/06_population_density.pkl', 'r') as fh:
    population = dill.load(fh)
population.head(1)

## Zika infection

with open('../pkl/03_infection_data_initial_import.pkl', 'r') as fh:
    infection = dill.load(fh)
infection.head(1)

infection = (infection[['report_date', 'location', 'value', 'data_field']]
             .rename(columns={'report_date':'date','value':'zika_cases'}))

for x in infection.zika_cases.iteritems():
    try:
        float(x[1])
    except:
        print(x)

infection.loc[2414, 'zika_cases'] = 0
infection.loc[2783, 'zika_cases'] = 0
infection.loc[5192, 'zika_cases'] = 0
infection['zika_cases'] = infection.zika_cases.fillna(0)
infection['zika_cases'] = infection.zika_cases.astype(int)

# def difference_mapper(df):
#     """Calculates differential entries and times.
#        Returns a dataframe."""
    
#     # Ensure these are sorted by date/time
#     df = df.sort_values('date')
    
#     # Calculate entry/time difference values and ensure none are zero/negative
#     if df.shape[0] > 1:
#         df['zika_cases'] = df.zika_cases.shift(-1) - df.zika_cases
    
#     return df

# infection2 = (infection.copy()
#  .groupby(['location', 'data_field'])
#  .apply(difference_mapper)
#  .sort_values('zika_cases', ascending=False)
# )

#infection = infection.loc[infection.zika_cases<90000]

infection = (infection
             .groupby(['location','date']).sum()
             .reset_index()
            )

infection.sort_values('zika_cases',ascending=False).head(20)

## Combine dataframes

model = pd.merge(weather_final,
                 infection,
                 on=['date','location'],
                 how='left')

model['zika_cases'] = model.zika_cases.fillna(0)

print model.shape, model.isnull().sum().max()

model = pd.merge(model,
                 population,
                 on='location', 
                 how='left')

print model.shape, model.isnull().sum().max()

model = pd.merge(model,
                 airport_distance,
                 on='location',
                 how='left')

print model.shape, model.isnull().sum().max()

model = pd.merge(model,
                 mosquito_distance,
                 on='location',
                 how='left')

print model.dropna().shape, model.isnull().sum().max()

save_it(model, '07_feature_engineering_and_cleaning')
