==> notebooks/09_scrape_GDP_and_PPP_data.ipynb <==
# Scrape world GDP and GDP based on purchase parity (PPP)

World GDP and purchase parity adjusted GDP from 2015 were scraped from [Knoema](http://knoema.com).

import pandas as pd
import numpy as np
import requests

import re
import time
from StringIO import StringIO

from selenium import webdriver
from selenium.webdriver.chrome.options import Options
from bs4 import BeautifulSoup

from csv_pkl_sql import save_it

## Scrape the GDP tables

driver = webdriver.Chrome()

driver.get('https://knoema.com/nwnfkne/world-gdp-ranking-2015-data-and-charts')
html = driver.page_source

# For some reason, scraping this page only works when 
# I save it to a file first and then read it back
with open('gdp.html', 'w') as fh:
    fh.write(html.encode('utf-8'))
    
with open('gdp.html','r') as fh:
    html = fh.read()
    
tables = pd.read_html(html)

! rm -rf gdp.html gdp_files

gdp = tables[2].rename(columns={'Unnamed: 0':'country','Units':'units'})
gdp.head(1)

gdp_ppp = tables[3].rename(columns={'Unnamed: 0':'country','Units':'units'})
gdp_ppp.head(1)

Select just the countries of interest.

locations = pd.read_pickle('../pkl/00_cleaned_city_names.pkl')
locations.head(1)

mask = gdp.country.isin(locations.country.unique())
gdp = gdp.loc[mask]
gdp.shape

mask = gdp_ppp.country.isin(locations.country.unique())
gdp_ppp = gdp_ppp.loc[mask]
gdp_ppp.shape

gdp_ppp.dtypes

gdp.dtypes

save_it(gdp, '09_GDP_table')

save_it(gdp_ppp, '09_GDP_PPP_table')
