==> notebooks/14_d3_files.ipynb <==
# Create D3 visualization 

This notebook creates a CSV file for use in a D3 visualization of Zika outbreaks over time. 

import pandas as pd
import numpy as np

from sklearn.metrics import confusion_matrix

%matplotlib inline
import matplotlib.pyplot as plt

## Load AdaBoost model and predict

fwf = pd.read_pickle('../pkl/10_class_balancing_fwf.pkl')

Ytrain = pd.read_pickle('../pkl/13_model_tuning_Ytrain_ada_inner_unbalanced.pkl')
Xtrain = pd.read_pickle('../pkl/13_model_tuning_Xtrain_ada_inner_unbalanced.pkl')

Ytest = pd.read_pickle('../pkl/13_model_tuning_Ytest_ada_inner_unbalanced.pkl')
Xtest = pd.read_pickle('../pkl/13_model_tuning_Xtest_ada_inner_unbalanced.pkl')

Ypred_test = pd.read_pickle('../pkl/13_model_tuning_Ypred_test_ada_inner_unbalanced.pkl')
Ypred_train = pd.read_pickle('../pkl/13_model_tuning_Ypred_train_ada_inner_unbalanced.pkl')

ada = pd.read_pickle('../pkl/13_model_tuning_ada_models_unbalanced.pkl')

# feat_cols = [x for x in features.columns if x not in ['date','location']]

print Xtest.shape, Ytest.shape

test = pd.merge(Xtest, Ytest.to_frame(), left_index=True, right_index=True)
test = pd.merge(test, Ypred_test.rename(columns={'zika_bool':'zika_pred'}), 
                left_index=True, right_index=True)
test = pd.merge(test, fwf[['location','date','cases_first_date','cases_total']], on=['location','date'], how='left')

print test.shape

print Xtrain.shape, Ytrain.shape

train = pd.merge(Xtrain, Ytrain.to_frame(), left_index=True, right_index=True)
train = pd.merge(train, Ypred_train.rename(columns={'zika_bool':'zika_pred'}), 
                left_index=True, right_index=True)
train = pd.merge(train, fwf[['location','date','cases_first_date','cases_total']], on=['location','date'], how='left')

print train.shape

data = pd.concat([test, train]).reset_index(drop=True)

data['cm_type'] = np.NaN

mask0 = (data.zika_bool.astype(bool) & data.zika_pred.astype(bool))
data.loc[mask0, 'cm_type'] = 0

mask1 = (data.zika_bool.astype(bool) & data.zika_pred.astype(bool).pipe(np.invert))
data.loc[mask1, 'cm_type'] = 1

mask2 = (data.zika_bool.astype(bool).pipe(np.invert) & data.zika_pred.astype(bool))
data.loc[mask2, 'cm_type'] = 3

mask3 = (data.zika_bool.astype(bool).pipe(np.invert) & data.zika_pred.astype(bool).pipe(np.invert))
data.loc[mask3, 'cm_type'] = 2

data['cm_type'] = data.cm_type.astype(np.int)

## Calculate confusion matrix

confusion_matrix(data['zika_bool'], data['zika_pred'])

data.cm_type.value_counts().sort_index()

## Get location coordinates

location_key_orig = pd.read_pickle('../pkl/00_cleaned_city_names.pkl')
location_key = pd.read_pickle('../pkl/01_latitude_longitude_google.pkl')

# cases = pd.read_pickle('../pkl/10_class_balancing_framework_a_first.pkl')
cases = pd.read_pickle('../pkl/03_infection_data_final.pkl')

print cases.shape

cases = cases[['date','location','zika_cases']].groupby(['location','date']).sum().reset_index()

print cases.shape

cases.head()

def join_names(df_row):

    subtype = df_row.location_type
    
    second_str = df_row[subtype]
    country_str = df_row.country
    
    return pd.Series({'country':country_str, 'city':second_str, 
                      'location':df_row.location})

location_key_orig = location_key_orig.apply(join_names, axis=1)

print location_key_orig.shape, location_key.shape, cases.shape

combined = pd.merge(location_key_orig.drop_duplicates(), location_key, on='location', how='left')

print combined.shape

combined = pd.merge(data[['location','date','zika_bool','zika_pred','cm_type']], 
                    combined, on='location', how='left')

print combined.shape

combined = pd.merge(cases, combined.drop(['date'],axis=1), on=['location'], how='right')

combined['zika_bool']

print combined.shape

combined.isnull().sum()

## Setup scaling of number of cases

(combined[['location','country','date','zika_cases']]
 .groupby('country')
 .sum()
  .sort_values('zika_cases',ascending=False))

# Something is probably wrong w/ scaling of cases in the Dominican Republic and Colombia
# as they shouldn't have more than Brazil. For now, scale these values down (see below)

mask = combined.country.isin(['Dominican Republic', 'Colombia'])
combined.loc[mask, 'zika_cases'] /= 10

tmp = combined.sort_values('zika_cases',ascending=False).head(1000)
tmp.cm_type.value_counts()

combined.rename(columns={'cm_type':'color'},inplace=True)

combined['zika_cases'] = (combined.zika_cases * 99./combined.zika_cases.max()) + 1
combined['location'] = combined.location.str.replace('-','_')

combined = (combined
 .sort_values('date')
 
 .rename(columns={'country':'state','location':'permalink',
                  'latitude':'lat', 'longitude':'long',
                  'date':'created_at'})
 [['city','state','permalink','created_at','lat','long','zika_cases','color']]
 .assign(permalink=combined.location + combined.date.dt.strftime("_%Y%m%d")))

# mask = (combined.state=='United States') & (combined.color==0)
tmp = combined[['city','state','lat','long']].drop_duplicates()

## Create CSV file for D3 visualization

# Remove some problematic data (non-ascii characters need to be cleaned)
mask = ((combined.state=='Colombia') & combined.city.str.startswith('Pe')).pipe(np.invert)
combined = combined.loc[mask]

(combined.to_csv('../map/zika_places.csv', 
            sep=',', 
            index=False, 
            date_format='%Y-%m-%d %H:%M:%S'))
! head ../map/zika_places.csv
